Moving to a more dependency-based setup
Previously certain parts of Helioviewer were packaged into containers.
We are moving to a more dependency-based approach where containers
support running the source code instead of packaging the source code
into the containers. We believe this will make it easier for anyone to
run, customize, and deploy Helioviewer in the future.